Modify: Modified Embedding module following to fixing vector fileds as List<Float> and taste, mood fileds as List<String> #5

+++ Embedding/src/main/kotlin/com/tastecompass/embedding/dto/EmbeddingResult.kt
@@ -1,6 +1,6 @@
 package com.tastecompass.embedding.dto
 
 data class EmbeddingResult(
-    val moodVector: List<Double>,
-    val tasteVector: List<Double>
+    val moodVector: List<Float>,
+    val tasteVector: List<Float>
 )
